Add explicit (separate) "docker-swarm" and "triton" docker-compose files
Since Triton does not yet support `docker network`, this is the best solution for making sure this repository can function appropriately in both environments (uses `docker network` for Swarm and `--link` for Triton).

Without this, Swarm will see `--link` and co-schedule all the containers, which defeats the whole purpose of the demonstration.